% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{attach_gene_symbol_from_entrez}
\alias{attach_gene_symbol_from_entrez}
\title{Add Gene Symbol from EntrezID}
\usage{
attach_gene_symbol_from_entrez(dat, species = "MUS")
}
\arguments{
\item{dat}{data frame with EnzrezID}

\item{species}{Either "MUS" or "HUM"}
}
\description{
Add Gene Symbol from EntrezID
}
\examples{
  tibble::tibble(
    ensembl_gene_id = c("ENSMUSG00000000001","ENSMUSG00000000003","ENSMUSG00000000028","ENSMUSG00000000031","ENSMUSG00000000037","ENSMUSG00000000049","ENSMUSG00000000056","ENSMUSG00000000058","ENSMUSG00000000078","ENSMUSG00000000085"),
  ) \%>\%
    rmyknife::ensembl_to_entrez() \%>\%
    rmyknife::attach_gene_symbol_from_entrez()
}
