% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{base_codename}
\alias{base_codename}
\title{Ubuntu codename of the image, used for the binary CRAN repository path}
\usage{
base_codename(labels)
}
\arguments{
\item{labels}{Named list of image labels}
}
\description{
Ubuntu codename of the image, used for the binary CRAN repository path
}
\keyword{internal}
