% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{deseq2_result_to_tibble}
\alias{deseq2_result_to_tibble}
\title{Convert a DESeq2 results object to a tibble}
\usage{
deseq2_result_to_tibble(dat, row_name = "ensembl_gene_id")
}
\arguments{
\item{dat}{DESeq2::results object}

\item{row_name}{The column name for the DESeq2 results row names}
}
\value{
tibble of DESeq2 result
}
\description{
Convert a DESeq2 results object to a tibble
}
\examples{
  deseq2_diff <- DESeq2::results(deseq2_ip_wt_vs_input_wt) \%>\%
    deseq2_result_to_tibble()
}
