% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{dt_datatable}
\alias{dt_datatable}
\title{Print standard DT::datatable with extensions for exporting the data}
\usage{
dt_datatable(dat, scroll_y)
}
\arguments{
\item{dat}{Dataframe to print}

\item{scroll_y}{Pixel size of data-table size in y}
}
\description{
Print standard DT::datatable with extensions for exporting the data
}
