% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{ensembl_to_entrez}
\alias{ensembl_to_entrez}
\title{Gets Entrez IDs and add it to dat}
\usage{
ensembl_to_entrez(
  dat,
  ensembl_id_name = "ensembl_gene_id",
  keep_only_rows_with_entrez = TRUE,
  drop_duplicates = TRUE,
  species = "MUS",
  ensembl = NULL,
  verbose = TRUE
)
}
\arguments{
\item{dat}{Data frame with ensembl identifier}

\item{ensembl_id_name}{Name of column containing the ensembl identifier}

\item{keep_only_rows_with_entrez}{Only keep rows for which entrez IDs could be found}

\item{drop_duplicates}{Often there is a n:1 relationship between Entrez-IDs and Ensembl-IDs. If this value is true, only keep the first hit}

\item{species}{Either MUS or HUM}

\item{ensembl}{Optional. If provided, use this biomaRt connection instead of OrgDb for conversion.}

\item{verbose}{Print summary statistics}
}
\description{
Gets Entrez IDs and add it to dat
}
\examples{
  tibble::tibble(
    ensembl_gene_id = c("ENSMUSG00000000001","ENSMUSG00000000003","ENSMUSG00000000028","ENSMUSG00000000031","ENSMUSG00000000037","ENSMUSG00000000049","ENSMUSG00000000056","ENSMUSG00000000058","ENSMUSG00000000078","ENSMUSG00000000085"),
  ) \%>\%
    rmyknife::ensembl_to_entrez()
}
