% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{ensure_lines}
\alias{ensure_lines}
\title{Append missing lines to a file, creating it when absent}
\usage{
ensure_lines(file, lines)
}
\arguments{
\item{file}{File to extend}

\item{lines}{Lines that must be present}
}
\description{
Append missing lines to a file, creating it when absent
}
\keyword{internal}
