% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{env_lock_path}
\alias{env_lock_path}
\title{Path to the environment lock file of a project}
\usage{
env_lock_path(path = ".")
}
\arguments{
\item{path}{Project root}
}
\description{
Path to the environment lock file of a project
}
\keyword{internal}
