% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_ensembl_host_from_version}
\alias{get_ensembl_host_from_version}
\title{Get ensembl host from ensembl_version without 'http://'
Example:
get_ensembl_host_from_version(92)
\link{1} "apr2018.archive.ensembl.org"}
\usage{
get_ensembl_host_from_version(ensembl_version)
}
\arguments{
\item{ensembl_version}{Integer of ensembl_version}
}
\description{
Get ensembl host from ensembl_version without 'http://'
Example:
get_ensembl_host_from_version(92)
\link{1} "apr2018.archive.ensembl.org"
}
