% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_gene_name_from_synonym}
\alias{get_gene_name_from_synonym}
\title{Get gene names from synonyms}
\usage{
get_gene_name_from_synonym(gene_name, species = "MUS", keep_missing = TRUE)
}
\arguments{
\item{gene_name}{Vector of gene names}

\item{species}{Define species. Default is "MUS"}

\item{keep_missing}{If we cannot identify a gene, still keep the original name?}
}
\description{
Get gene names from synonyms
}
