% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_genes_of_goterm_godb}
\alias{get_genes_of_goterm_godb}
\title{Get genes associated with GO-term as vector based on GO.db}
\usage{
get_genes_of_goterm_godb(go_accession, species = "MUS", verbose = TRUE,
  memoised = TRUE)
}
\arguments{
\item{go_accession}{ID of GO term}

\item{species}{Define species, either "HUM" or "MUS"}

\item{verbose}{Print summary statistic of the query}

\item{memoised}{Use memoised function (search for local caches of given parameters)}
}
\description{
Get genes associated with GO-term as vector based on GO.db
}
\examples{
  get_genes_of_goterm_godb("GO:1900746")
}
