% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_genes_of_goterm_godb_helper}
\alias{get_genes_of_goterm_godb_helper}
\title{Get genes associated with GO-term as vector based on GO.db}
\usage{
get_genes_of_goterm_godb_helper(go_accession, species = "MUS",
  verbose = TRUE)
}
\arguments{
\item{go_accession}{See get_genes_of_goterm_godb}

\item{species}{See get_genes_of_goterm_godb}

\item{verbose}{See get_genes_of_goterm_godb}
}
\description{
Get genes associated with GO-term as vector based on GO.db
}
\examples{
  get_genes_of_goterm_godb_helper("GO:1900746")
}
