% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_genes_of_goterm_helper}
\alias{get_genes_of_goterm_helper}
\title{Get genes associated with GO-term based on BiomaRt
This will get the genes of the selected GO-term and
\emph{all} child-terms.}
\usage{
get_genes_of_goterm_helper(go_accession, ensembl)
}
\arguments{
\item{go_accession}{ID of GO term}

\item{ensembl}{Biomart connection}
}
\description{
Get genes associated with GO-term based on BiomaRt
This will get the genes of the selected GO-term and
\emph{all} child-terms.
}
