% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_goterm_name_from_id}
\alias{get_goterm_name_from_id}
\title{Get GO name.}
\usage{
get_goterm_name_from_id(go_accession, ensembl)
}
\arguments{
\item{go_accession}{GO term id, e.g. "GO:0032680"}

\item{ensembl}{Ensembl dataset object}
}
\description{
Get GO name.
}
\examples{
   get_goterm_name_from_id(go_accession = "GO:0032680", ensembl = rmyknife::get_ensembl_dataset_from_version(94, "mmusculus_gene_ensembl"))
}
