% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_goterms_gene_list}
\alias{get_goterms_gene_list}
\title{Get a list of all GO-terms and genes from Biomart}
\usage{
get_goterms_gene_list(ensembl = get_ensembl_dataset_from_version())
}
\arguments{
\item{ensembl}{Mart for query}
}
\value{
tibble with gene-goterm association
}
\description{
Get a list of all GO-terms and genes from Biomart
}
