% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_memoised}
\alias{get_memoised}
\title{Wrapper function for getting Memoised function
Returns memoised function if option rmyknife.use_memoise is true}
\usage{
get_memoised(func)
}
\description{
Wrapper function for getting Memoised function
Returns memoised function if option rmyknife.use_memoise is true
}
