% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_orgdb_for_species}
\alias{get_orgdb_for_species}
\title{Get orgdb for species as a string}
\usage{
get_orgdb_for_species(species = "MUS")
}
\arguments{
\item{species}{Either MUS or HUM}
}
\value{
OrgDB database string
}
\description{
Get orgdb for species as a string
}
