% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_promotor_sequence}
\alias{get_promotor_sequence}
\title{Get promotor sequence upstream of a gene flank}
\usage{
get_promotor_sequence(ensembl_gene_ids,
  ensembl = get_ensembl_dataset_from_version(97),
  upstream_bases = 1000)
}
\arguments{
\item{ensembl_gene_ids}{genes to get sequence from}

\item{ensembl}{biomaRt connection}

\item{upstream_bases}{How many bases upstream of the gene start to return}
}
\value{
tibble with columns gene_flank and ensembl_gene_id
}
\description{
Get promotor sequence upstream of a gene flank
}
\examples{
   get_promotor_sequence(
     ensembl_gene_ids = c("ENSMUSG00000102693", "ENSMUSG00000064842", "ENSMUSG00000102851"),
     ensembl = rmyknife::get_ensembl_dataset_from_version(94)
   )
   get_promotor_sequence(
     ensembl_gene_ids = c("ENSG00000140505", "ENSG00000205358", "ENSG00000125144", "ENSG00000198417", "ENSG00000205364", "ENSG00000169715", "ENSG00000187193", "ENSG00000125148"),
     ensembl = rmyknife::get_ensembl_dataset_from_version(97, species = "HUM")
   )
}
