% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_promotor_sequence_tibble}
\alias{get_promotor_sequence_tibble}
\title{Wrapper function for get_promotor_sequence. Attaches the result to a data frame}
\usage{
get_promotor_sequence_tibble(dat, ensembl_id_var = "ensembl_gene_id",
  ensembl = get_ensembl_dataset_from_version(97),
  upstream_bases = 1000)
}
\arguments{
\item{dat}{tibble with ensembl_gene_id variable}

\item{ensembl_id_var}{ensembl id variable. Default is "ensembl_gene_id"}

\item{ensembl}{... from get_promotor_sequence}

\item{upstream_bases}{... from get_promotor_sequence}
}
\value{
original tibble with column gene_flank added
}
\description{
Wrapper function for get_promotor_sequence. Attaches the result to a data frame
}
\examples{
  tibble::tibble(EnsemblIDs = c("ENSMUSG00000102693", "ENSMUSG00000064842", "ENSMUSG00000102851")) \%>\%
    get_promotor_sequence_add_tibble(
      ensembl_id_var = "EnsemblIDs"
    )
}
