% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrnaseq.R
\name{get_seurat_clustering}
\alias{get_seurat_clustering}
\title{Seurat clustering based on https://satijalab.org/seurat/v3.1/pbmc3k_tutorial.html}
\usage{
get_seurat_clustering(
  seurat_data,
  max_dimension_used_for_clustering = 10,
  cluster_resolution = 0.8
)
}
\arguments{
\item{seurat_data}{Seurat dataset object}

\item{max_dimension_used_for_clustering}{Determines how many principal components are used for clustering}

\item{cluster_resolution}{See Seurat::FindClusters resolution parameter. "resolution: Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities."}
}
\description{
Seurat clustering based on https://satijalab.org/seurat/v3.1/pbmc3k_tutorial.html
}
\examples{
cellatlas::get_campbell_count_data() \%>\%
  # transform into format that can be read by seurat
  dplyr::select(GeneID, CellID, count) \%>\%
  tidyr::spread(CellID, count) \%>\%
  tibble::column_to_rownames("GeneID") \%>\%
  # Initialize the Seurat object with the raw (non-normalized data)
  Seurat::CreateSeuratObject(
    counts = .,
    min.cells = 3,
    min.features = 200,
    project = "campbell"
  ) \%>\%
  # Get seurat clustering
  get_seurat_clustering()
}
