% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_tempfile_from_url}
\alias{get_tempfile_from_url}
\title{Downloads the file in the URL to a temp file and returns it's path}
\usage{
get_tempfile_from_url(url)
}
\arguments{
\item{url}{File url}
}
\value{
path to temp file
}
\description{
Downloads the file in the URL to a temp file and returns it's path
}
\examples{
   get_tempfile_from_url("https://ars.els-cdn.com/content/image/1-s2.0-S0092867419300571-mmc2.xlsx")
}
