% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_uniprot}
\alias{get_uniprot}
\title{Helper function for constructing uniprot API call}
\usage{
get_uniprot(search_string, species = "MUS", verbose = TRUE)
}
\arguments{
\item{search_string}{Search Uniprot database for genes matching this search}

\item{species}{Either "HUM" or "MUS". Leave empty for pulling all species}

\item{verbose}{Print status messages}
}
\description{
Helper function for constructing uniprot API call
}
