% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment-verify.R
\name{git_run}
\alias{git_run}
\title{Run git in a project and return its output}
\usage{
git_run(path, args)
}
\arguments{
\item{path}{Project root}

\item{args}{Arguments to git}
}
\description{
Run git in a project and return its output
}
\keyword{internal}
