% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{has_conflicting_repos}
\alias{has_conflicting_repos}
\title{Does an .Rprofile set repositories in a way that would fight the pin block?}
\usage{
has_conflicting_repos(lines)
}
\arguments{
\item{lines}{Lines of the .Rprofile}
}
\description{
Does an .Rprofile set repositories in a way that would fight the pin block?
}
\keyword{internal}
