% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{has_repo_pins}
\alias{has_repo_pins}
\title{Does an .Rprofile already pin dated repositories?}
\usage{
has_repo_pins(lines)
}
\arguments{
\item{lines}{Lines of the .Rprofile}
}
\description{
Does an .Rprofile already pin dated repositories?
}
\keyword{internal}
