% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{lockfile_info}
\alias{lockfile_info}
\title{Read version and repository pins out of an existing renv lockfile}
\usage{
lockfile_info(path = ".")
}
\arguments{
\item{path}{Project root}
}
\description{
Read version and repository pins out of an existing renv lockfile
}
\keyword{internal}
