% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{make}
\alias{make}
\title{Make targets and load them into the global enrivonment}
\usage{
make(envir = parent.frame(), load_to_environment = TRUE, workers = 10)
}
\arguments{
\item{envir}{The environment to load the targets into}

\item{load_to_environment}{Whether to load the targets into the environment after building}

\item{workers}{Number of workers to use for make (requires crew package if > 1)}
}
\value{
Number of outdated targets
}
\description{
Make targets and load them into the global enrivonment
}
