% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{pin_block}
\alias{pin_block}
\title{The repository pin block written into \code{.Rprofile}}
\usage{
pin_block(snapshot_date, bioc_version, codename = "noble")
}
\arguments{
\item{snapshot_date}{CRAN snapshot date}

\item{bioc_version}{Bioconductor release}

\item{codename}{Ubuntu codename for binary packages}
}
\description{
The repository pin block written into \code{.Rprofile}
}
\keyword{internal}
