% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{pinned_repos}
\alias{pinned_repos}
\title{Pinned repository vector}
\usage{
pinned_repos(snapshot_date, bioc_version, codename = "noble")
}
\arguments{
\item{snapshot_date}{CRAN snapshot date}

\item{bioc_version}{Bioconductor release}

\item{codename}{Ubuntu codename for binary packages}
}
\description{
Pinned repository vector
}
\keyword{internal}
