% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_volcano}
\alias{plot_volcano}
\title{Volcano plot using data frame containing columns \code{log2FoldChange} and `padj``}
\usage{
plot_volcano(dat, min_padj = 0.05, min_log2fc = 1, label_top_n = 10)
}
\arguments{
\item{dat}{data frame containing columns \code{log2FoldChange} and \code{padj}}

\item{min_padj}{Minimum p-value cutoff}

\item{min_log2fc}{Minimum log2fc cutoff}

\item{label_top_n}{Top-n genes to label}
}
\description{
Volcano plot using data frame containing columns \code{log2FoldChange} and `padj``
}
