% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{print_params}
\alias{print_params}
\title{Print params list with knitr}
\usage{
print_params(pars, omit = c())
}
\arguments{
\item{pars}{Parameter list to print}

\item{omit}{Character vector with names of parameters that are not printed}
}
\description{
Print params list with knitr
}
\examples{
  print_params(rmd_params, "counts")
}
