% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment-verify.R
\name{print_verify}
\alias{print_verify}
\title{Print a verification report}
\usage{
print_verify(report)
}
\arguments{
\item{report}{Tibble as returned by \code{\link[=project_verify]{project_verify()}}}
}
\description{
Print a verification report
}
\keyword{internal}
