% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{read_cellranger_feature_bc_matrix}
\alias{read_cellranger_feature_bc_matrix}
\title{Read Cell Ranger matrix and return a tibble
Code adapted from https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/output/matrices#r-load-mat}
\usage{
read_cellranger_feature_bc_matrix(matrix_dir, tidy = TRUE)
}
\arguments{
\item{matrix_dir}{Path of files "matrix.mtx.gz", "features.tsv.gz" and "barcodes.tsv.gz"}

\item{tidy}{Return tidy tibble}
}
\value{
tibble of resulting matrix
}
\description{
Read Cell Ranger matrix and return a tibble
Code adapted from https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/output/matrices#r-load-mat
}
