% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrnaseq.R
\name{read_cellranger_feature_bc_matrix_h5}
\alias{read_cellranger_feature_bc_matrix_h5}
\title{Read Cell Ranger matrix and return a tibble
Code adapted from https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/output/matrices#r-load-mat}
\usage{
read_cellranger_feature_bc_matrix_h5(h5_path, use_names = FALSE, tidy = TRUE)
}
\arguments{
\item{h5_path}{Path of file}

\item{use_names}{Keep original ensembl IDs or use clear gene names}

\item{tidy}{Return tidy tibble}
}
\value{
tibble of resulting matrix
}
\description{
Read Cell Ranger matrix and return a tibble
Code adapted from https://support.10xgenomics.com/single-cell-gene-expression/software/pipelines/latest/output/matrices#r-load-mat
}
