% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{read_cufflinks}
\alias{read_cufflinks}
\title{Read in data from cufflinks tsv}
\usage{
read_cufflinks(path)
}
\arguments{
\item{path}{Path to cufflinks tsv file}
}
\value{
Tibble of cufflinks data
}
\description{
Read in data from cufflinks tsv
}
