% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{read_env_lock}
\alias{read_env_lock}
\title{Read the environment lock of a project}
\usage{
read_env_lock(path = ".")
}
\arguments{
\item{path}{Project root}
}
\value{
The environment lock as a list
}
\description{
Read the environment lock of a project
}
