% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{rprofile_subdir_template}
\alias{rprofile_subdir_template}
\title{\code{.Rprofile} for a subdirectory, delegating to the project root}
\usage{
rprofile_subdir_template()
}
\description{
\code{.Rprofile} for a subdirectory, delegating to the project root
}
\keyword{internal}
