% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{rprofile_template}
\alias{rprofile_template}
\title{Full \code{.Rprofile} template for a project that does not have one yet}
\usage{
rprofile_template(snapshot_date, bioc_version, codename = "noble")
}
\arguments{
\item{snapshot_date}{CRAN snapshot date}

\item{bioc_version}{Bioconductor release}

\item{codename}{Ubuntu codename for binary packages}
}
\description{
Full \code{.Rprofile} template for a project that does not have one yet
}
\keyword{internal}
