% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment-verify.R
\name{status_if}
\alias{status_if}
\title{Status of a condition, as "ok" or a chosen failure level}
\usage{
status_if(ok, level = "fail")
}
\arguments{
\item{ok}{Logical condition}

\item{level}{Status to use when \code{ok} is FALSE}
}
\description{
Status of a condition, as "ok" or a chosen failure level
}
\keyword{internal}
