% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{status_message}
\alias{status_message}
\title{Emit a tagged progress message}
\usage{
status_message(status, ...)
}
\arguments{
\item{status}{One of "ok", "warn", "fail" or "info"}

\item{...}{Parts of the message, pasted together}
}
\description{
Emit a tagged progress message
}
\keyword{internal}
