% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment-verify.R
\name{verify_images}
\alias{verify_images}
\title{Check that every recorded image is present and unchanged}
\usage{
verify_images(env_lock, deep)
}
\arguments{
\item{env_lock}{Environment lock}

\item{deep}{Whether to verify checksums}
}
\description{
Check that every recorded image is present and unchanged
}
\keyword{internal}
