% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{write_env_lock}
\alias{write_env_lock}
\title{Write the environment lock}
\usage{
write_env_lock(env_lock, path)
}
\arguments{
\item{env_lock}{Environment lock list}

\item{path}{Destination file}
}
\description{
Write the environment lock
}
\keyword{internal}
