% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{attach_ensembl_gene_id_from_entrez_id}
\alias{attach_ensembl_gene_id_from_entrez_id}
\title{Attach Biomart Gene identifier from entrez id}
\usage{
attach_ensembl_gene_id_from_entrez_id(
  dat,
  entrez_id_var = "entrez_id",
  ensembl = get_ensembl_dataset_from_version(100, species = "MUS"),
  verbose = TRUE
)
}
\arguments{
\item{dat}{input data frame containing entrez id's}

\item{entrez_id_var}{Entrez id column in character type.}

\item{ensembl}{ensembl database connection object}

\item{verbose}{Print summary statistics to check for 1:1 or 1:N mappings}
}
\value{
Tibble of dat with attached biomaRt variables
}
\description{
Attach Biomart Gene identifier from entrez id
}
\examples{
   tibble(
     entrez_gene_ids = c("4496", "4494", "4495", "1544"),
     my_id = 1:4
   ) \%>\%
   rmyknife::attach_ensembl_gene_id_from_entrez_id(entrez_id_var = "entrez_gene_ids")

}
