% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_genes_of_goterm}
\alias{get_genes_of_goterm}
\title{Get genes associated with GO-term based on BiomaRt}
\usage{
get_genes_of_goterm(go_accession, ensembl, verbose = TRUE)
}
\arguments{
\item{ensembl}{Biomart connection}

\item{go_id}{ID of GO term}
}
\value{
Tibble with all genes associated with GO-term
}
\description{
Get genes associated with GO-term based on BiomaRt
}
\examples{
ion_transport_genes <- get_genes_of_goterm(go_accession = "GO:0006811", ensembl = rmyknife::get_ensembl_dataset_from_version(94, "mmusculus_gene_ensembl"))
}
