% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biomart.R
\name{get_goterm_name_from_id}
\alias{get_goterm_name_from_id}
\title{DEPRECATED. Get GO name using GO.db}
\usage{
get_goterm_name_from_id(go_accession)
}
\arguments{
\item{go_accession}{GO term id, e.g. "GO:0032680"}
}
\value{
GO-term name
}
\description{
DEPRECATED. Get GO name using GO.db
}
\examples{
   get_goterm_name_from_id(go_accession = "GO:0032680")
}
