% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{get_seurat_clustering}
\alias{get_seurat_clustering}
\title{Seurat clustering based on https://satijalab.org/seurat/v3.1/pbmc3k_tutorial.html}
\usage{
get_seurat_clustering(seurat_data,
  max_dimension_used_for_clustering = 10, cluster_resolution = 0.8)
}
\arguments{
\item{seurat_data}{Seurat dataset object}

\item{max_dimension_used_for_clustering}{Determines how many principal components are used for clustering}

\item{cluster_resolution}{See Seurat::FindClusters resolution parameter. "resolution: Value of the resolution parameter, use a value above (below) 1.0 if you want to obtain a larger (smaller) number of communities."}
}
\description{
Seurat clustering based on https://satijalab.org/seurat/v3.1/pbmc3k_tutorial.html
}
