% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{make}
\alias{make}
\title{Make targets and load them into the global enrivonment}
\usage{
make(envir = parent.frame())
}
\arguments{
\item{envir}{The environment to load the targets into}
}
\description{
Make targets and load them into the global enrivonment
}
