% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{makefile_template}
\alias{makefile_template}
\title{Stub makefile for a project that does not have one}
\usage{
makefile_template(image, bind)
}
\arguments{
\item{image}{Path to the primary image}

\item{bind}{Bind mounts}
}
\description{
Stub makefile for a project that does not have one
}
\keyword{internal}
