% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment-backup.R
\name{restore_instructions}
\alias{restore_instructions}
\title{Instructions written into every backup}
\usage{
restore_instructions(env_lock, archive, members, lfs = FALSE)
}
\arguments{
\item{env_lock}{Environment lock}

\item{archive}{Archive file name}

\item{members}{Files contained in the archive}

\item{lfs}{Whether the project tracks files with git-lfs}
}
\description{
Instructions written into every backup
}
\keyword{internal}
