% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{write_xls}
\alias{write_xls}
\title{WriteXLS::WriteXLS() function with sane default parameter}
\usage{
write_xls(dat, ExcelFileName, SheetNames = filename)
}
\arguments{
\item{dat}{data frame to export}

\item{ExcelFileName}{see WriteXLS::WriteXLS()}

\item{SheetNames}{see WriteXLS::WriteXLS()}
}
\description{
WriteXLS::WriteXLS() function with sane default parameter
}
